Updated packages

+++ src/main/java/com/hotel/Hotel_Reservation_Management/repository/RoomRepository.java
@@ -5,6 +5,7 @@ import com.hotel.Hotel_Reservation_Management.entity.Room;
 import com.hotel.Hotel_Reservation_Management.enums.RoomStatus;
 import com.hotel.Hotel_Reservation_Management.enums.RoomType;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface RoomRepository extends JpaRepository<Room, Long> {
@@ -15,5 +16,5 @@ public interface RoomRepository extends JpaRepository<Room, Long> {
 
     List<Room> findByRoomType(RoomType roomType);
 
-    List<Room> findByPriceLessThanEqual(Double price);
+    List<Room> findByPriceLessThanEqual(BigDecimal price);
 }
